Разделил код на модули
