+++ js/main.mjs
@@ -1,0 +1,4 @@
+import {generatePhotosArray} from "./post.js"
+const photosArray = generatePhotosArray();
+console.log(photosArray);
+
