+++ js/util.js
@@ -1,0 +1,26 @@
+// Функция для генерации случайного числа от min до max
+function getRandomNumber(min, max) {
+    return Math.floor(Math.random() * (max - min + 1)) + min;
+  }
+
+    // Функция для генерации случайного комментария
+    function generateRandomComment() {
+        const avatars = ['img/avatar-1.svg', 'img/avatar-2.svg', 'img/avatar-3.svg', 'img/avatar-4.svg', 'img/avatar-5.svg', 'img/avatar-6.svg'];
+        const messages = [
+          'Всё отлично!',
+          'В целом всё неплохо. Но не всё.',
+          'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+          'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+          'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+          'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+        ];
+      
+        return {
+          id: getRandomNumber(1, 10000),
+          avatar: avatars[getRandomNumber(1, 6)],
+          message: messages[getRandomNumber(1, 6)],
+          name: 'User' + '-' + getRandomNumber(1, 100)
+        };
+      }
+  
+export{getRandomNumber,generateRandomComment}
